# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Captora Photography CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /shoots:
    get:
      operationId: listShoots
      tags: [shoots]
      summary: List all shoots
      parameters:
        - name: clientId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of shoots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Shoot"
    post:
      operationId: createShoot
      tags: [shoots]
      summary: Create a shoot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShootInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shoot"
  /shoots/{id}:
    get:
      operationId: getShoot
      tags: [shoots]
      summary: Get shoot by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Shoot details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShootDetail"
    patch:
      operationId: updateShoot
      tags: [shoots]
      summary: Update a shoot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShootUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shoot"
    delete:
      operationId: deleteShoot
      tags: [shoots]
      summary: Delete a shoot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /shoots/{id}/status:
    patch:
      operationId: updateShootStatus
      tags: [shoots]
      summary: Update shoot status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShootStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shoot"
  /shoots/{id}/team:
    get:
      operationId: getShootTeam
      tags: [shoots]
      summary: Get team assignments for a shoot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Team assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ShootTeamMember"
    post:
      operationId: assignTeamToShoot
      tags: [shoots]
      summary: Assign team members to shoot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShootTeamInput"
      responses:
        "200":
          description: Updated team
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ShootTeamMember"

  # PAYMENTS (CLIENT)
...
  /staff:
    get:
      operationId: listStaff
      tags: [staff]
      summary: List all staff/employees
      responses:
        "200":
          description: List of staff
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffMember"
    post:
      operationId: createStaffMember
      tags: [staff]
      summary: Add a staff member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
  /staff/{id}:
    get:
      operationId: getStaffMember
      tags: [staff]
      summary: Get staff member by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Staff member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffDetail"
    patch:
      operationId: updateStaffMember
      tags: [staff]
      summary: Update a staff member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffMember"
    delete:
      operationId: deleteStaffMember
      tags: [staff]
      summary: Delete a staff member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /staff/{id}/salary:
    get:
      operationId: listSalaryRecords
      tags: [staff]
      summary: List salary records for a staff member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Salary records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SalaryRecord"
    post:
      operationId: addSalaryRecord
      tags: [staff]
      summary: Add or update monthly salary record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SalaryRecordInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalaryRecord"

  # EXPENSES
...
components:
  schemas:
...
    Shoot:
      type: object
      required: [id, clientId, shootDate, status, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: ["string", "null"]
        shootDate:
          type: string
        shootTime:
          type: ["string", "null"]
        venue:
          type: ["string", "null"]
        functions:
          type: array
          items:
            type: string
        status:
          type: string
        specialInstructions:
          type: ["string", "null"]
        deliveryNotes:
          type: ["string", "null"]
        createdAt:
          type: string
    ShootDetail:
      allOf:
        - $ref: "#/components/schemas/Shoot"
        - type: object
          properties:
            team:
              type: array
              items:
                $ref: "#/components/schemas/ShootTeamMember"
            equipmentChecklist:
              type: array
              items:
                type: string
    ShootInput:
      type: object
      required: [clientId, shootDate]
      properties:
        clientId:
          type: integer
        shootDate:
          type: string
        shootTime:
          type: string
        venue:
          type: string
        functions:
          type: array
          items:
            type: string
        specialInstructions:
          type: string
        equipmentChecklist:
          type: array
          items:
            type: string
    ShootUpdate:
      type: object
      properties:
        shootDate:
          type: string
        shootTime:
          type: string
        venue:
          type: string
        functions:
          type: array
          items:
            type: string
        specialInstructions:
          type: string
        deliveryNotes:
          type: string
        equipmentChecklist:
          type: array
          items:
            type: string
    ShootStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
        deliveryNotes:
          type: string
    ShootTeamMember:
      type: object
      required: [id, shootId, memberId, memberType]
      properties:
        id:
          type: integer
        shootId:
          type: integer
        memberId:
          type: integer
        memberType:
          type: string
        memberName:
          type: ["string", "null"]
        role:
          type: ["string", "null"]
        callTime:
          type: ["string", "null"]
    ShootTeamInput:
      type: object
      required: [members]
      properties:
        members:
          type: array
          items:
            type: object
            required: [memberId, memberType]
            properties:
              memberId:
                type: integer
              memberType:
                type: string
              callTime:
                type: string
...
    StaffMember:
      type: object
      required: [id, name, role, monthlySalary, joiningDate, phone, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
        monthlySalary:
          type: number
        joiningDate:
          type: string
        phone:
          type: string
        createdAt:
          type: string
    StaffDetail:
      allOf:
        - $ref: "#/components/schemas/StaffMember"
        - type: object
          properties:
            salaryRecords:
              type: array
              items:
                $ref: "#/components/schemas/SalaryRecord"
    StaffInput:
      type: object
      required: [name, role, monthlySalary, joiningDate, phone]
      properties:
        name:
          type: string
        role:
          type: string
        monthlySalary:
          type: number
        joiningDate:
          type: string
        phone:
          type: string
    StaffUpdate:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        monthlySalary:
          type: number
        phone:
          type: string
    SalaryRecord:
      type: object
      required: [id, staffId, month, year, amount, status, createdAt]
      properties:
        id:
          type: integer
        staffId:
          type: integer
        month:
          type: integer
        year:
          type: integer
        amount:
          type: number
        status:
          type: string
        paidDate:
          type: ["string", "null"]
        createdAt:
          type: string
    SalaryRecordInput:
      type: object
      required: [month, year, amount, status]
      properties:
        month:
          type: integer
        year:
          type: integer
        amount:
          type: number
        status:
          type: string
        paidDate:
          type: string
...