# lib/api-zod/src/generated/types/staffInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Captora Photography CRM API
 * OpenAPI spec version: 0.1.0
 */

export interface StaffInput {
  name: string;
  role: string;
  monthlySalary: number;
  joiningDate: string;
  phone: string;
}